perf: lazy-load DynamicIcon module to reduce main bundle size
Extract DynamicIcon to a separate module that's only loaded when
action-tracker buttons have icons configured. This reduces the main
bundle from 821KB to 724KB (~97KB savings) by deferring the 184KB
dynamicIconImports map until needed.
